backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymongo
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
firebase-admin==6.4.0
requests
python-dotenv
pandas
numpy
scikit-learn
joblib
geojson
eth-account
web3
motor
loguru
pydantic[email]